[ëFILE] 이메일 기ë인증¥ 구현현

+++ src/main/java/com/example/gms/product/repository/ProductRepository.java
@@ -2,6 +2,5 @@ package com.example.gms.product.repository;
 
 import com.example.gms.product.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
